Assignments/HW8/Part2_Agentic_System/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
kafka-python-ng==2.2.2
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
